Added the ability to have holes inside of both the polygon and multipolygon.
The GeoJSON spec says the first path in a polygon is the exterior ring and all
other paths are holes.

The google implementation seems to allow holes if their direction (clockwise
or counterclokwise) is reversed from the initial path. This seems like it
could be a problem with some complex or complicated polygons.